linter: move sideEffectFree function to solver package (#611)
